Updated licence and introduced vnrId
The vnfr id in the recoveryAction is needed so that when the FMS receive and event from a vnfr it can associate the event to the recovery action. For example when a switch to standby action is finished the fms can remove the recovery action in progress relative to that vnfr.